Minor refactor of timestamp and username formatting
